Added default option to include value at `t_stop`
Without that, `write_to_cml` misses the last row of the CML with the
largest timestamp if no `t_stop` is supplied, since `t_stop` is set
exactly to the largest timestamp. In case files should be split up,
we have however exclude the `t_stop` since it will be `t_start` for the
next file.