Added option to pass `max_distance` for IDW, but not sure if this is what we want
The idea was that this produces NaNs further than `max_distance` away from the CMLs, but it, of course, also impacts the interpolation closer to a CML since all CMLs further away are not included as neighbors and hence the results may look a bit "eddgy". Further evaluation will be needed.