improved dataframe reindex aggregation function and added test (#31)
The `ComlinkGridInterpolator` now support a new keyword argument `resample_to_new_index` which can be used to aggregate the CML data onto arbitrary time steps before doing the interpolation (which is done for the individual time steps). This functionality is handy when interpolated fields have to be generated with the same rainfall accumulation periods as the
german met service hourly radar product RADOLAN, which aggregates from, e.g. 00:50 till 01:50, labeling it 01:50.

Additional info:

* improved dataframe reindex aggregation function and added test

* added basic unit test for aggregation on new index

* To keep the speed acceptable, `get_dataframe_for_cml_variable` now first builds a large DataFrame for all CMLs using their individual DatetimeIndices. This might produce a huge DataFrame when doing the concat because the time stamps of all CMLs are a little different, resulting in a lot of NaNs in the concated DataFrame