Speed up of `calc_coverage_mask` and added example application (#35)
This improves the speed of  `calc_coverage_mask` for large numbers of CMLs. With the old version it would have taken an hour or more to do it for the RADOLAN grid and 4000 CMLs. Now it takes a minute. The speed up is not noticeable with small numbers of CMLs, though.

The notebook was update to show how a CML coverage mask is calculated and applied.