Fix interpolator bug when having different amount of NaN in consecutive input data
The problem was that self.x and self.y have been set with x and y that have not been subseted to the values where z has no NaNs. Because the interpolator tries to reuse the KDtree if x and y do not change this led to the problem that the interpolator did not create a new tree when z had more NaNs then when used for initializing the KDtree.