add xarray wrapper for waa_lejnse
Note that it was important here to make sure the L_km is a scalar value because otherwise the wrapper fails during the iteration when calling another function, not sure if it was calc_R_from_A or if its was somewhere else. But L_km should be a scalar values for one CML for sure anyway.