Fixed IDW code based on what I did in the RADOLAN-CML codebase and added a new test
The test would fail with the old IDW code because I messed up the thing
with the weights and the numba loop that I added when adopting the code
from stackoverflow.

Unfortunately the fixes to the IDW code do not resolve the issue with
numba and testing with `export NUMBA_DISABLE_JIT=1`. The error is still
there when I try this out locally with pytest, see
https://github.com/pycomlink/pycomlink/issues/83