Move and extend grid intersection code
Grid intersection code was moved to new grid_intersection module where
it makes much more sense.

A function for deriving grid intersection weights for several CMLs and
storing them as sparse arrays in an xarray.DataArray was added. Test for
this is also there.

Next step is to remove the old grid_intersection code. Not clear what to
do with the validator classes, though. They are not really used now,
    since I will add a function for getting the grid intersection time
    series with sparse tensordot in the next commits.