Disable build of PDF and epub on readthedocs
We need this because the PDF build fails and we do not care about the PDF or epub build. Note that this only seems to affect the builds done of the master branch and not the PR builds of the docs which apparently only build the html version of the docs (which makes sense for a PR docs build which is used for debugging). Hence we did not see this problem occur in #142.

I hope this solved the issue with the current main readthedocs build.